airflow_setup/requirements.txt
apache-airflow[postgres,gcp]==1.10.7
# black==19.10b0
# pylint==2.4.4
# docker==4.1.0